requirements.txt
matplotlib~=3.3.4
numpy~=1.19.5
Pillow~=8.1.0
torch~=1.7.1
torchvision~=0.8.2
tensorboard
future
tqdm~=4.56.2
opencv-python~=4.5.1.48
imageio~=2.9.0
nibabel~=3.2.1
torchsummary~=1.5.1
h5py~=2.10.0
MedPy~=0.4.0
albumentations~=0.5.2
scikit-image~=0.18.1
nnunet~=1.6.6
argparse~=1.4.0
batchgenerators~=0.21
sklearn~=0.0
scikit-learn~=0.24.1